pipeline:refactor: splitting steps as classes
