+++ eva-pipeline/src/main/java/embl/ebi/variation/eva/pipeline/jobs/VariantConfiguration.java
@@ -16,11 +16,10 @@
 package embl.ebi.variation.eva.pipeline.jobs;
 
 import embl.ebi.variation.eva.pipeline.listeners.JobParametersListener;
-import org.opencb.biodata.models.variant.VariantSource;
-import org.opencb.datastore.core.QueryOptions;
-import org.opencb.opencga.storage.core.variant.VariantStorageManager;
-import org.opencb.opencga.storage.core.variant.adaptors.VariantDBAdaptor;
-import org.opencb.opencga.storage.core.variant.stats.VariantStatisticsManager;
+import embl.ebi.variation.eva.pipeline.steps.VariantsLoad;
+import embl.ebi.variation.eva.pipeline.steps.VariantsStatsCreate;
+import embl.ebi.variation.eva.pipeline.steps.VariantsStatsLoad;
+import embl.ebi.variation.eva.pipeline.steps.VariantsTransform;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.batch.core.*;
@@ -28,24 +27,14 @@ import org.springframework.batch.core.configuration.annotation.JobBuilderFactory
 import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
 import org.springframework.batch.core.job.builder.JobBuilder;
 import org.springframework.batch.core.launch.JobLauncher;
-import org.springframework.batch.core.scope.context.ChunkContext;
 import org.springframework.batch.core.step.builder.StepBuilder;
 import org.springframework.batch.core.step.builder.TaskletStepBuilder;
-import org.springframework.batch.core.step.tasklet.Tasklet;
-import org.springframework.batch.repeat.RepeatStatus;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.*;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Set;
 
-import org.opencb.datastore.core.ObjectMap;
-import org.opencb.opencga.storage.core.StorageManagerFactory;
 import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.batch.core.launch.support.RunIdIncrementer;
 import org.springframework.core.env.Environment;
@@ -58,10 +47,6 @@ public class VariantConfiguration {
     public static final String jobName = "variantJob";
     public static final String loadJobName = "variantLoadJob";
     public static final String statsJobName = "variantStatsJob";
-//    public static final String annotJobName = "variantAnnotationJob";
-    public static final String SKIP_LOAD = "skipLoad";
-    public static final String SKIP_STATS_CREATE = "skipStatsCreate";
-    public static final String SKIP_STATS_LOAD = "skipStatsLoad";
 
     @Autowired
     private JobBuilderFactory jobBuilderFactory;
@@ -122,31 +107,7 @@ public class VariantConfiguration {
 
     public Step transform() {
         StepBuilder step1 = stepBuilderFactory.get("transform");
-        TaskletStepBuilder tasklet = step1.tasklet(new Tasklet() {
-            @Override
-            public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
-                JobParameters parameters = chunkContext.getStepContext().getStepExecution().getJobParameters();
-                ObjectMap variantOptions = listener.getVariantOptions();
-
-                URI outdirUri = createUri(parameters.getString("outputDir"));
-                URI nextFileUri = createUri(parameters.getString("input"));
-                URI pedigreeUri = parameters.getString("pedigree") != null ? createUri(parameters.getString("pedigree")) : null;
-
-                logger.info("transform file " + parameters.getString("input") + " to " + parameters.getString("outputDir"));
-
-                logger.info("Extract variants '{}'", nextFileUri);
-                VariantStorageManager variantStorageManager = StorageManagerFactory.getVariantStorageManager();
-                variantStorageManager.extract(nextFileUri, outdirUri, variantOptions);
-
-                logger.info("PreTransform variants '{}'", nextFileUri);
-                variantStorageManager.preTransform(nextFileUri, variantOptions);
-                logger.info("Transform variants '{}'", nextFileUri);
-                variantStorageManager.transform(nextFileUri, pedigreeUri, outdirUri, variantOptions);
-                logger.info("PostTransform variants '{}'", nextFileUri);
-                variantStorageManager.postTransform(nextFileUri, variantOptions);
-                return RepeatStatus.FINISHED;
-            }
-        });
+        TaskletStepBuilder tasklet = step1.tasklet(new VariantsTransform(listener));
 
         // true: every job execution will do this step, even if this step is already COMPLETED
         // false: if the job was aborted and is relaunched, this step will NOT be done again
@@ -157,37 +118,7 @@ public class VariantConfiguration {
 
     public Step load() {
         StepBuilder step1 = stepBuilderFactory.get("load");
-        TaskletStepBuilder tasklet = step1.tasklet(new Tasklet() {
-            @Override
-            public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
-                JobParameters parameters = chunkContext.getStepContext().getStepExecution().getJobParameters();
-                ObjectMap variantOptions = listener.getVariantOptions();
-
-                if (Boolean.parseBoolean(parameters.getString(SKIP_LOAD, "false"))) {
-                    logger.info("skipping load step, requested " + SKIP_LOAD + "=" + parameters.getString(SKIP_LOAD));
-                } else {
-                    VariantStorageManager variantStorageManager = StorageManagerFactory.getVariantStorageManager();// TODO add mongo
-                    URI outdirUri = createUri(parameters.getString("outputDir"));
-                    URI nextFileUri = createUri(parameters.getString("input"));
-                    URI pedigreeUri = parameters.getString("pedigree") != null ? createUri(parameters.getString("pedigree")) : null;
-                    Path output = Paths.get(outdirUri.getPath());
-                    Path input = Paths.get(nextFileUri.getPath());
-                    Path outputVariantJsonFile = output.resolve(input.getFileName().toString() + ".variants.json" + parameters.getString("compressExtension"));
-//                outputFileJsonFile = output.resolve(input.getFileName().toString() + ".file.json" + config.compressExtension);
-                    URI transformedVariantsUri = outdirUri.resolve(outputVariantJsonFile.getFileName().toString());
-
-
-                    logger.info("-- PreLoad variants -- {}", nextFileUri);
-                    variantStorageManager.preLoad(transformedVariantsUri, outdirUri, variantOptions);
-                    logger.info("-- Load variants -- {}", nextFileUri);
-                    variantStorageManager.load(transformedVariantsUri, variantOptions);
-//                logger.info("-- PostLoad variants -- {}", nextFileUri);
-//                variantStorageManager.postLoad(transformedVariantsUri, outdirUri, variantOptions);
-                }
-
-                return RepeatStatus.FINISHED;
-            }
-        });
+        TaskletStepBuilder tasklet = step1.tasklet(new VariantsLoad(listener));
 
         // true: every job execution will do this step, even if this step is already COMPLETED
         // false: if the job was aborted and is relaunched, this step will NOT be done again
@@ -197,33 +128,7 @@ public class VariantConfiguration {
 
     public Step statsCreate() {
         StepBuilder step1 = stepBuilderFactory.get("statsCreate");
-        TaskletStepBuilder tasklet = step1.tasklet(new Tasklet() {
-            @Override
-            public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
-//                HashMap<String, Set<String>> samples = new HashMap<>(); // TODO fill properly. if this is null overwrite will take on
-//                samples.put("SOME", new HashSet<>(Arrays.asList("HG00096", "HG00097")));
-                JobParameters parameters = chunkContext.getStepContext().getStepExecution().getJobParameters();
-
-                if (Boolean.parseBoolean(parameters.getString(SKIP_STATS_CREATE, "false"))) {
-                    logger.info("skipping stats creation step, requested " + SKIP_STATS_CREATE + "=" + parameters.getString(SKIP_STATS_CREATE));
-                } else {
-                    ObjectMap variantOptions = listener.getVariantOptions();
-                    VariantStorageManager variantStorageManager = StorageManagerFactory.getVariantStorageManager();
-                    VariantSource variantSource = variantOptions.get(VariantStorageManager.VARIANT_SOURCE, VariantSource.class);
-                    VariantDBAdaptor dbAdaptor = variantStorageManager.getDBAdaptor(variantOptions.getString("dbName"), variantOptions);
-                    URI outdirUri = createUri(parameters.getString("outputDir"));
-                    URI statsOutputUri = outdirUri.resolve(VariantStorageManager.buildFilename(variantSource));
-
-                    VariantStatisticsManager variantStatisticsManager = new VariantStatisticsManager();
-                    QueryOptions statsOptions = new QueryOptions(variantOptions);
-
-                    // actual stats creation
-                    variantStatisticsManager.createStats(dbAdaptor, statsOutputUri, null, statsOptions);    // TODO allow subset of samples
-                }
-
-                return RepeatStatus.FINISHED;
-            }
-        });
+        TaskletStepBuilder tasklet = step1.tasklet(new VariantsStatsCreate(listener));
 
         // true: every job execution will do this step, even if this step is already COMPLETED
         // false: if the job was aborted and is relaunched, this step will NOT be done again
@@ -233,44 +138,11 @@ public class VariantConfiguration {
 
     public Step statsLoad() {
         StepBuilder step1 = stepBuilderFactory.get("statsLoad");
-        TaskletStepBuilder tasklet = step1.tasklet(new Tasklet() {
-            @Override
-            public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
-
-                JobParameters parameters = chunkContext.getStepContext().getStepExecution().getJobParameters();
-
-                if (Boolean.parseBoolean(parameters.getString(SKIP_STATS_LOAD, "false"))) {
-                    logger.info("skipping stats loading");
-                } else {
-                    VariantStorageManager variantStorageManager = StorageManagerFactory.getVariantStorageManager();
-                    ObjectMap variantOptions = listener.getVariantOptions();
-                    QueryOptions statsOptions = new QueryOptions(variantOptions);
-                    VariantStatisticsManager variantStatisticsManager = new VariantStatisticsManager();
-                    VariantDBAdaptor dbAdaptor = variantStorageManager.getDBAdaptor(variantOptions.getString("dbName"), variantOptions);
-                    URI outdirUri = createUri(parameters.getString("outputDir"));
-                    VariantSource variantSource = variantOptions.get(VariantStorageManager.VARIANT_SOURCE, VariantSource.class);
-                    URI statsOutputUri = outdirUri.resolve(VariantStorageManager.buildFilename(variantSource));
-
-                    // actual stats load
-                    variantStatisticsManager.loadStats(dbAdaptor, statsOutputUri, statsOptions);
-                }
-
-                return RepeatStatus.FINISHED;
-            }
-        });
+        TaskletStepBuilder tasklet = step1.tasklet(new VariantsStatsLoad(listener));
 
         // true: every job execution will do this step, even if this step is already COMPLETED
         // false: if the job was aborted and is relaunched, this step will NOT be done again
         tasklet.allowStartIfComplete(false);
         return tasklet.build();
     }
-
-    public static URI createUri(String input) throws URISyntaxException {
-        URI sourceUri = new URI(null, input, null);
-        if (sourceUri.getScheme() == null || sourceUri.getScheme().isEmpty()) {
-            sourceUri = Paths.get(input).toUri();
-        }
-        return sourceUri;
-    }
-
 }
